+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [9.2.0] - (16.08.2024)
+
+## Changes
+
+- add generator mode Off/On/Auto control from home app as a extra tiles 
+- arf profile refactor
+- cleanup
+
 ## [9.1.3] - (16.08.2024)
 
 ## Changes